+++ cypress/integration/other/Rules.spec.ts
@@ -1,6 +1,7 @@
 describe('Rules', () => {
   beforeEach(() => {
     cy.visit('/');
+    cy.get('[data-testid="close-warning"]').click();
     cy.get('[data-testid=options]').click();
     cy.contains('Rules').click();
   });